+++ lib/entities/reorderable_entity.dart
@@ -1,17 +1,47 @@
 import 'package:flutter/cupertino.dart';
 
+/// Represents [child] with some extra information.
+///
+/// When animating all children in this package, this entity is important
+/// to have access to all needed values of [child].
+///
+/// With this entity, it is possible to know where the [child] is
+/// positioned and which order that [child] has inside all children.
+///
+/// Also the current state of [child] is added as information: [isBuilding],
+/// [isNew] and [hasSwappedOrder].
 class ReorderableEntity {
+  /// Represents this entity
   final Widget child;
+
+  /// Describes [size] of [child].
   final Size size;
 
+  /// Describes the original orderId before it was updated.
   final int originalOrderId;
+
+  /// Describes the updated orderId when it was updated.
   final int updatedOrderId;
 
+  /// Describes the original [Offset] before it was updated.
   final Offset originalOffset;
+
+  /// Describes the updated [Offset] when it was updated.
   final Offset updatedOffset;
 
+  /// Usually means that the [child] has a new position that is still unknown.
+  ///
+  /// If [isBuilding] is true, then it is possible, that the [Offset] and
+  /// orderId will be updated.
   final bool isBuilding;
+
+  /// The [Offset] can already be known but this is still a flag to know, that [child] didn't exist before.
   final bool isNew;
+
+  /// Is true, when this [child] only changed the position with another child.
+  ///
+  /// This is only true, when the changed position has nothing to do with
+  /// another added or removed child.
   final bool hasSwappedOrder;
 
   const ReorderableEntity({
@@ -26,6 +56,7 @@ class ReorderableEntity {
     this.hasSwappedOrder = false,
   });
 
+  /// Overrides all parameters of this entity and returns the updated [ReorderableEntity].
   ReorderableEntity copyWith({
     Offset? originalOffset,
     Offset? updatedOffset,
@@ -49,5 +80,6 @@ class ReorderableEntity {
         hasSwappedOrder: hasSwappedOrder ?? this.hasSwappedOrder,
       );
 
+  /// Getting faster access to the hashCode of the child's key.
   int get keyHashCode => child.key.hashCode;
 }